web-ui/requirements-web.txt
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
pg2ch==0.1.1
